Casted return to float, formated file using black

+++ Pyrado/pyrado/tasks/reward_functions.py
@@ -428,4 +428,4 @@ class QCartPoleSwingUpRewFcn(RewFcn):
             raise pyrado.TypeErr(given=err_a, expected_type=np.ndarray)
 
         # Reward should be roughly between [0, 1]
-        return self.factor*(1-np.abs(err_s[1]/np.pi)**2) + (1-self.factor)*(np.abs(err_s[0])) 
+        return float(self.factor * (1 - np.abs(err_s[1] / np.pi) ** 2) + (1 - self.factor) * (np.abs(err_s[0])))
